add post info

+++ src/routes/post.rs
@@ -27,20 +27,20 @@ pub fn get_post_routes(client: Arc<Client>) -> Router {
                 }
             }
         })).route("/posts/:id", get({
-            move |path: axum::extract::Path<i32>, extension: Extension<Arc<Client>>| {
+            move |path: axum::extract::Path<i64>, extension: Extension<Arc<Client>>| {
                 async move {
                     get_post(path, extension).await
                 }
             }
         })).route("/posts/:id", delete({
-            move |path: axum::extract::Path<i32>, extension: Extension<Arc<Client>>| {
+            move |path: axum::extract::Path<i64>, extension: Extension<Arc<Client>>| {
                 async move {
                     delete_post(path, extension).await
                 }
             }
         })).route("/posts/:id", put({
             let client_clone = client.clone(); // Clone for this closure
-            move |path: axum::extract::Path<i32>, Json(post): Json<Post>| {
+            move |path: axum::extract::Path<i64>, Json(post): Json<Post>| {
                 let extension = Extension(client_clone);
                 async move {
                     update_post(path, extension, Json(post)).await
