+++ src/routes/user.rs
@@ -33,14 +33,14 @@ pub fn get_user_routes(client: Arc<Client>) -> Router {
                 }
             }
         })).route("/users/:id", delete({
-            move |path: axum::extract::Path<i32>, extension: Extension<Arc<Client>>| {
+            move |path: axum::extract::Path<i64>, extension: Extension<Arc<Client>>| {
                 async move {
                     delete_user(path, extension).await
                 }
             }
         })).route("/users/:id", put({
             let client_clone = client.clone(); // Clone for this closure
-            move |path: axum::extract::Path<i32>, Json(user): Json<User>| {
+            move |path: axum::extract::Path<i64>, Json(user): Json<User>| {
                 let extension = Extension(client_clone);
                 let name = user.name.clone();
                 let email = user.email.clone();
